requirements.txt
# UrbanAlign 实验依赖库

# 数据处理
pandas>=1.3.0
numpy>=1.21.0

# 机器学习
scikit-learn>=1.0.0
scipy>=1.7.0

# TrueSkill评分系统
trueskill>=0.4.5

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0

# 进度条
tqdm>=4.62.0

# API请求
requests>=2.26.0

# Environment variable management
python-dotenv>=0.19.0

# 可选: CLIP特征提取(如需自己提取)
# torch>=1.10.0
# transformers>=4.20.0
# ftfy>=6.0.0

# JSON处理(Python内置,无需安装)
# json

# Base64编码(Python内置)
# base64
